weather/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
openpyxl==3.1.2
pandas==2.0.1
python-decouple==3.7